feat: separates out demo components into own components
